# openapi.yaml
openapi: 3.0.3
info:
  title: ChoirLoop API
  description: |
    API for ChoirLoop - A browser-based choir practice tool that allows choir members 
    to practice specific sections from MIDI files with individual voice control.
  version: 1.0.0
  contact:
    name: ChoirLoop Support
    url: https://choirloop.rosakehlchen.de
servers:
  - url: https://choirloop.rosakehlchen.de/api
    description: Production server
  - url: http://localhost:8080/api
    description: Local development server
...
paths:
...
  /songs:
    get:
      tags:
        - songs
      summary: List all songs
      description: Retrieve a list of all available songs
      responses:
        '200':
          description: List of songs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  songs:
                    type: array
                    items:
                      $ref: '#/components/schemas/SongSummary'
    
    post:
      tags:
        - songs
      summary: Create a new song
      description: Create a new song entry (without files)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: "Ave Maria"
                description:
                  type: string
                  example: "Choir practice for Sunday service"
      responses:
        '201':
          description: Song created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  song:
                    $ref: '#/components/schemas/Song'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /songs/{id}:
    get:
      tags:
        - songs
      summary: Get song details
      description: Retrieve detailed information about a specific song
      parameters:
        - $ref: '#/components/parameters/SongId'
      responses:
        '200':
          description: Song details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  song:
                    $ref: '#/components/schemas/Song'
        '404':
          description: Song not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - songs
      summary: Update song
      description: Update song metadata
      parameters:
        - $ref: '#/components/parameters/SongId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Song updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  song:
                    $ref: '#/components/schemas/Song'
        '404':
          description: Song not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - songs
      summary: Delete song
      description: Delete a song and all its associated files
      parameters:
        - $ref: '#/components/parameters/SongId'
      responses:
        '200':
          description: Song deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Song deleted successfully
        '404':
          description: Song not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    SongId:
      name: id
      in: path
      required: true
      description: Unique identifier of the song (UUID)
      schema:
        type: string
        format: uuid
        example: "ecb12847-edae-4c3e-b743-d4337c63cb5b"
...
  schemas:
    SongSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "ecb12847-edae-4c3e-b743-d4337c63cb5b"
        title:
          type: string
          example: "Ave Maria"
        description:
          type: string
          example: "Practice for Sunday service"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-08T22:20:28+00:00"
    Song:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "ecb12847-edae-4c3e-b743-d4337c63cb5b"
        title:
          type: string
          example: "Ave Maria"
        description:
          type: string
          example: "Practice for Sunday service"
        midi_file:
          type: string
          nullable: true
          example: "song.mid"
        score_file:
          type: string
          nullable: true
          example: "score.xml"
        voices:
          type: array
          items:
            $ref: '#/components/schemas/Voice'
        practice_sections:
          type: array
          items:
            $ref: '#/components/schemas/PracticeSection'
        created_at:
          type: string
          format: date-time
          example: "2025-11-08T20:00:00+00:00"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-08T22:20:28+00:00"
    Voice:
      type: object
      properties:
        track_number:
          type: integer
          description: MIDI track number (0-indexed)
          example: 0
        name:
          type: string
          description: Voice name (e.g., Soprano, Alto, Tenor, Bass)
          example: "Soprano 1"
        channel:
          type: integer
          nullable: true
          description: MIDI channel number
          example: 1
        note_count:
          type: integer
          description: Number of notes in this track
          example: 245
    PracticeSection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        label:
          type: string
          description: Human-readable label for the section
          example: "Chorus"
        start_measure:
          type: integer
          minimum: 1
          description: Starting measure number
          example: 5
        start_beat:
          type: integer
          minimum: 1
          description: Starting beat within the measure
          example: 1
        end_measure:
          type: integer
          minimum: 1
          description: Ending measure number
          example: 12
        end_beat:
          type: integer
          minimum: 1
          description: Ending beat within the measure
          example: 4
        created_at:
          type: string
          format: date-time
          example: "2025-11-08T20:00:00+00:00"
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Song not found"
        message:
          type: string
          description: Detailed error description
          example: "The requested song does not exist"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Currently not implemented. Future versions may require authentication.

# Storage Implementation Notes
...